Guida

+++ weblogicmvc/webapp/controller/AirportController.php
@@ -8,8 +8,37 @@ use ArmoredCore\WebObjects\View;
 
 class AirportController extends BaseController
 {
+    private function utilizadorIsValid()
+    {
+        // verifica se o utilizador atual tem acesso às funções deste controlador
+
+        if (!Session::has("Authentication")) {
+            MensagensHelper::setError("Nao tem permissoes para  aceder à area do administrador");
+            Redirect::toRoute('home/login');
+            return false;
+        }
+
+        $userAuth = Session::get('Authentication');
+
+        if ($userAuth["perfil"] != 'administrador') {
+            MensagensHelper::setError("Nao tem permissoes para  aceder à area do administrador");
+            Redirect::toRoute('home/login');
+            return false;
+        }
+
+        return true;
+    }
+
+    // =========================================================
+    //                          AEROPORTOS
+    // =========================================================
+
     public function criar()
     {
+        if (!$this->utilizadorIsValid()) {
+            return;
+        }
+
         // obtem o aeroporto do POST do form
         $airport = Post::getAll();
 
@@ -17,10 +46,10 @@ class AirportController extends BaseController
         $exist = Airport::all(array('conditions' => array('nome = ? AND pais = ?', $airport['nome'], $airport['pais'])));
 
         if ($exist) {
-            Session::set("error-message", "Já existe o aeroporto com o nome: " . $airport['nome']);
+            MensagensHelper::setError( "Já existe o aeroporto com o nome: " . $airport['nome']);
         } else {
             Airport::create($airport);
-            Session::set("success-message", "Aeroporto " . $airport['nome'] . " criado com sucesso!");
+            MensagensHelper::setSuccess("Aeroporto " . $airport['nome'] . " criado com sucesso");
         }
 
         Redirect::toRoute('admin/aeroportos');
@@ -28,6 +57,10 @@ class AirportController extends BaseController
 
     public function editar($idaeroporto)
     {
+        if (!$this->utilizadorIsValid()) {
+            return;
+        }
+
         $airport = Airport::find([$idaeroporto]);
 
         if (!is_null($airport)) {
@@ -38,27 +71,32 @@ class AirportController extends BaseController
 
     public function atualizar($idaeroporto)
     {
+        if (!$this->utilizadorIsValid()) {
+            return;
+        }
+
         $airport = Airport::find([$idaeroporto]);
         $airport->update_attributes(Post::getAll());
+        $airport->save();
 
-        if ($airport->is_valid()) {
-            $airport->save();
-            Redirect::toRoute('admin/aeroportos');
-        } else {
-            // TODO: show errors
-            //Redirect::flashToRoute('aeroporto/editar', ['airport' => $airport]);
-        }
+        MensagensHelper::setSuccess("Aeroporto atualizado com sucesso");
+
+        Redirect::toRoute('admin/aeroportos');
     }
 
     public function eliminar($idaeroporto)
     {
+        if (!$this->utilizadorIsValid()) {
+            return;
+        }
+
         $airport = Airport::find([$idaeroporto]);
 
         if (is_null($airport)) {
-            Session::set("error-message", "Não foi possível eliminar o Aeroporto com o id " . $idaeroporto);
+            MensagensHelper::setError( "Não foi possível eliminar o Aeroporto com o id " . $idaeroporto);
         } else {
             $airport->delete();
-            Session::set("success-message", "Aeroporto com o id " . $idaeroporto . " eliminado com sucesso!");
+            MensagensHelper::setSuccess("Aeroporto com o nome " . $airport->nome . " eliminado com sucesso");
         }
 
         Redirect::toRoute('admin/aeroportos');
